Create a second IR between Pieces and output string.
Before this PR, CodeWriter eagerly built a complete string of formatted code using a StringBuffer. This is simple and works OK, but it's a little slow when separate formatting gets involved. Since a subtree of a Piece tree might be formatted separately, it means we often build a string using a StringBuffer only to then append that result to some surrounding StringBuffer, recursively.

This PR introduces a slightly more abstract representation for "formatted code but not quite a single string". It lets us compose separately formatted subtrees by just adding a single existing object to a list.

The performance gain isn't huge, but is measurable:

```
Benchmark (tall)                fastest   median  slowest  average  baseline
-----------------------------  --------  -------  -------  -------  --------
block                             0.063    0.065    0.114    0.069    101.0%
chain                             0.605    0.618    0.645    0.619    103.6%
collection                        0.159    0.163    0.175    0.164    101.4%
collection_large                  0.856    0.896    3.295    0.997    100.1%
conditional                       0.065    0.067    0.086    0.069     99.7%
curry                             0.547    0.565    0.596    0.565    106.8%
ffi                               0.144    0.150    0.170    0.151    110.0%
flutter_popup_menu_test           0.262    0.272    0.295    0.272    103.4%
flutter_scrollbar_test            0.123    0.125    0.142    0.128    103.7%
function_call                     1.287    1.339    1.474    1.342    101.1%
infix_large                       0.619    0.638    0.669    0.642    107.2%
infix_small                       0.159    0.171    0.719    0.228    105.5%
interpolation                     0.090    0.091    0.114    0.093    100.9%
large                             3.456    3.499    3.537    3.499    101.7%
top_level                         0.141    0.145    0.168    0.147    102.2%
```

And when run on a large repo:

```
Current formatter     8.536 ========================================
Optimized             8.226 ======================================
Old formatter         4.441 ====================
The current formatter is 47.97% slower than the old formatter.
The optimized is 3.77% faster than the current formatter.
The optimized is 46.01% slower than the old formatter.
The optimization gets the formatter 7.57% of the way to the old one.
```

The other motivation for this change is that I'm starting to work on being able to opt a region of code out of formatting and I suspect (but am not sure yet) that this representation will make it easier to slot the existing formatted code into the output as it's being lowered to a string.