Move selection and other random tests into a new directory.
There are only a few selection test files, and there were a couple of
tiles in top_level/ that didn't really have to do with top-level code
and were just in there for lack of a better home.

I'm about to add tests for disabling formatting, so I figured it makes
sense to have an "other" directory for all of these tests. This commit
just creates that directory and moves the existing tests there.