Delete unreachable `default` clause. (#1547)
The Dart analyzer will soon be changed so that if the `default` clause
of a `switch` statement is determined to be unreachable by the
exhaustiveness checker, a new warning of type
`unreachable_switch_default` will be issued. This parallels the
behavior of the existing `unreachable_switch_case` warning, which is
issued whenever a `case` clause of a `switch` statement is determined
to be unreachable.

This PR deletes an unreachable `default` clause from `dart_style` now,
to avoid a spurious warning when the analyzer change lands.