Handle imports with "as" and "if" clauses in either order. (#1555)
PR #1550 fixed the bug where if you had both clauses, it would output
them in the *wrong* order. But it turns out that the parser today allows
them in *either* order even though the language spec doesn't.

Since I've seen a handful of cases like this in the wild, instead of
just failing to format, simply preserve the clause order that the user
wrote.