Add code for working with analysis_options.yaml files.
This can find, read, and merge them. It handles `include` inside the
file.

It works using an abstraction over the file system so that (in theory at
least), this code could be harvested and made reusable by other tools
that need to work with analysis_options.yaml files but don't want to
work directly with files on disk.

It also includes an implementation of that abstraction using dart:io.

Right now, this is all just private inside dart_style, but is organized
so that it would be easy to pull out into a separate package later if we
want.