Apply class modifiers to all the classes. (#1575)
Now that we've bumped the major version and can make breaking changes,
I went ahead and locked down all of the classes. It mostly matters for
the public ones, but I went ahead and applied it to the internal ones
too just to see if it makes things easier to maintain.