Better formatting for long type parameter bounds. (#1576)
Better formatting for long type parameter bounds.

The tall style was already doing a better job on #1568 than the old
short style does. (I don't know why it leaves an overly long line there.
It seems like a bug but perhaps a nasty subtle one in the solver.)

But it still didn't allow splitting before `extends` in a bound which
could leave long lines in cases where the type parameter and bound name
are both quite long. Rare, but we may as well do a better job.

So this introduces a new piece that allows splitting before `extends`
if needed.

Fix #1568.