Rename "--line-length" option to "--page-width".
The formatter has always been inconsistent about how to refer to the
limit that it tries to fit code in. The option name (which is older
than when I started working on the formatter) is "--line-length". But
the public API and all of the internal code uses "page width".

This renames that option to "--page-width". The old name is still
supported (but hidden in help) for backwards compatibility.

In practice, this option isn't used often, so this likely won't affect
many users.