Allow a factory constructor with metadata before it to split. (#1580)
Prior to this PR, if there was metadata before a constructor, it would
be attached to the "header" piece of the constructor -- the leading
keywords and name -- instead of the entire constructor piece.

That in turn meant that the newline after the metadata would lead the
formatter to think a split had occurred inside the header itself,
instead of before the entire constructor. That would in turn confuse
the solver and it ended up trying to disallow splitting anywhere in the
constructor, even if that led to an overflowing line.

This fixes that so that the metadata is attached to the entire
constructor piece, not just the nested header part.