Clean up README and TODO comments. (#1592)
Clean up README and TODO comments.

Updated the example in the README to one that uses the new style. I
also took the opportunity to find a more interesting example and
generally rewrote things based on how the formatter is used today.

When we were in the middle of the rewrite, I used "TODO(tall)" and
"TODO(perf)" comments to track things I wanted to look into before I
considered the rewrite "done". Now that the rewrite is basically good
enough, I went through those. If I think the TODO is no longer worth
doing at all, I just deleted it. Otherwise, I turned it into a general
"TODO(rnystrom)" for us to keep in mind at some unspecified time.