Simplify ChainPiece.
We were still passing in allowNewlines to _formatCall() even though that
parameter was no longer used once Piece got the allowChildInState() API.
So I removed that.

Then I noticed that the switch in _formatCall() could be more cleanly
handled by having the cases in format() calculate separate directly. So
I did that.

That left _formatCall() not actually doing anything useful, so I got rid
of it.

There are no behavioral change, it's just a code clean-up.