Make "// @dart=" version comments affect the applied style.
The formatter uses an incoming language version to control how the
formatted code is parsed. The langauge version is also used to
determine whether you get the short or tall style.

Prior to this PR, a "// @dart=" version comment in the formatted code
would affect the language version the code is parsed at. But it wouldn't
use that comment to determine whether you get the short or tall style.
This fixes that.

Fix #1599.