Better indentation for property access as binary operator operand.
While I was at it, I figured out a more systematic way to hoist leading comments out from the subexpressions containing the token that the comments precede. That fixes this bug and also lets us get rid of many of the one-off places where we were manually hoisting comments out of some subexpressions.

Surprisingly, even though this touches visitNode() which is a fairly hot method, it doesn't seem to measurably harm performance.