Add an optional configuration to preserve trailing commas.
In the old short formatter, when a delimited construct had a trailing comma, it had two effects:

1.  The construct would be formatted "Flutter style" where the closing bracket would be moved to the next line and the contents indented +2, as opposed to the earlier style where the bracket stays on the same line as the last element and the elements are indented +4).

2.  The construct is forced to split even if it otherwise wouldn't.

The new formatter always uses a Flutter-like style for commas-separated constructors, so trailing commas are no longer needed for point 1.

When it does split a construct, it always adds a trailing comma. This means that point 2 breaks reversibility. Because of that, and because many users don't want to decide for every construct whether it should split or not, the new tall formatter no longer does point 2. It will remove a trailing comma and collapse a construct if it decides to.

However, some users rely heavily on that feature and strongly prefer control over forcing argument lists and other constructs to split.

This PR restores that functionality with an opt-in configuration:

```
formatter:
  trailing_commas: preserve
```

When enabled, if a construct has a trailing comma, the formatter will always split it. Users that preferred that behavior from the old formatter can enable the option and continue to work that way.

This does not mean we are generally moving in the direction of a more configurable formatter. This was a behavior that the formatter already supported and we removed it to the detriment of some users' experience. We're restoring that existing behavior for those users who want it.

Fix #1652.