Start adding support for "shape" parent constraints.
Currently, a parent can only indicate whether a child can contain newlines or not when determining whether a solution is invalid. This
works fine for most pieces, but isn't sophisticated enough to allow fixing #1465 and #1466 where we want to say that the child can have newlines but only in certain styles.

This is a step towards supporting those more sophisticated constraints. It doesn't change the underlying semantics at all, but it replaces the basic boolean newlines yes/no, with a Set<Shape>.