Use indentation merging to handle flattened indentation in assignments.
It looks nice if we align binary operands when the first operand is on its own line, which is often the case after `=`, `:`, and `=>`:

```dart
variable =
    operand +
    another;
```

This is also true for adjacent strings and conditional expressions.

Previously, we had some somewhat hacky code that looked at the AST node surrounding an expression to determine if it should indent itself or not. This removes that in favor of using the new merging ability of Indent. We use a separate Indent type for assignment and infix operators and then merge those two together.

This simplifies a bunch of things and lends itself to more uniform formatting since every AST node that lowers to a piece using Indent.infix will get this behavior. In fact, this commit highlighted that `is` and `as` expressions currently *aren't* formatted the same way as other binary operators, which I think was just an oversight.

This commit still formats them differently, but I'll probably change that later. For now, it just leaves TODOs.

This commit does change the formatter behavior very slightly, but only in obscure edge cases.